moved all test/database.py logic over to conftest.py because pytest will automatically put that into the scope of any tests in that folder.
