+++ tests/test_users.py
@@ -1,19 +1,6 @@
 from app import schemas
-from .database import client, session
-import pytest
 from jose import jwt
 from config import settings
-
-
-@pytest.fixture
-def test_user(client):
-    user_data = {"email":"voldemort@gmail.com", "password":"123"}
-    res = client.post('/users/', json=user_data)
-    assert res.status_code == 201
-    # print(res.json())
-    new_user = res.json()
-    new_user['password'] = user_data['password']
-    return new_user
 
 
 def test_create_users(client):
